Create helper classes instead of touching interface.
